Window 1 Date Draw Adjustments
Date is now drawn underneath the mouse centered on the mouse position.
The date is now displayed correctly as the offsets from the left and
right sides were adjusted. The date is now only drawn in window 1 if the
mouse is actually over that window. This function can be reused for the
Window 2 as well if we want the date to be able to be moused over there
as well, which I am assuming will happen. I also want to have a similar
function written on Window 2 only where the data points you are
currently mousing over are displayed in raw text sort of like how the
date is in Window 1, it follows the mouse, but I want it to actually be
following the mouse's Y axis position as well. This way, the user can
see the exact values for the individual data points there. Data on
demand I guess.