Restricted Data selection to strictly the bounds of Window 1
This restriction prevents the rectangle from being drawn off the window
1 plane and also allows for future changes to include selecting and
manipulating data in other windows with the mouse without affecting what
is drawn in window 1. Also, if the mouse is pressed off Window 1 and
then dragged onto window 1 to start making a selection, the initial
selection value will start at the left most value on Window 1. Another
option is to have the initial startX value set to wherever the mouse
enters the window 1 area, however I am going to stick to the left most
value for now for simplicities sake.